Updated documentation. Moved files to organize them differently
